+++ CHANGELOG.md
@@ -5,6 +5,7 @@ unreleased
 
 ## Feat
 - `user delete`で`username`を指定できるように
+- `note delete`で`visibility`を指定できるように
 
 ## Fix